+++ src/main/java/com/saborexpress/saborexpress/service/ClienteService.java
@@ -6,17 +6,17 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 @Service
-public interface  ClienteService {
+public interface ClienteService {
 
     List<Cliente> findAll();
 
     List<Cliente> findByNome(final String nome);
 
-    Optional<Cliente> findById(final Long id);
+    Optional<Cliente> findById(final Integer id);
 
     Cliente save(final Cliente cliente);
 
-    Optional<Cliente> update(final Long id, final Cliente clienteAtualizado);
+    Optional<Cliente> update(final Integer id, final Cliente clienteAtualizado);
 
-    void delete(final Long id);
+    void delete(final Integer id);
 }
